feat(extrinsics): move ExtrinsicOpts to a separate file
